build: Simplify tree-sitter-all-languages crate
