+++ docker-images/syntax-highlighter/crates/scip-treesitter-cli/Cargo.toml
@@ -7,8 +7,6 @@ edition = "2021"
 name = "scip-treesitter"
 
 [dependencies]
-lazy_static = "1.0"
-paste = "1.0.6"
 assert_cmd = "2.0.12"
 predicates = "3.0.4"
 insta = "1.34.0"
@@ -19,6 +17,8 @@ protobuf = { workspace = true }
 scip = { workspace = true }
 anyhow = { workspace = true }
 clap = { workspace = true }
+lazy_static = { workspace = true }
+paste = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 string-interner = { workspace = true }
