+++ docker-images/syntax-highlighter/crates/syntax-analysis/Cargo.toml
@@ -9,6 +9,8 @@ edition = "2021"
 anyhow.workspace = true
 clap.workspace = true
 itertools.workspace = true
+lazy_static.workspace = true
+paste.workspace = true
 protobuf.workspace = true
 rocket.workspace = true
 scip.workspace = true
