+++ docker-images/syntax-highlighter/crates/syntax-analysis/src/bin/scip-local-nav.rs
@@ -3,7 +3,7 @@ use std::{fs, path::Path};
 use clap::Parser;
 use scip::{types::Document, write_message_to_file};
 use syntax_analysis::{languages::LocalConfiguration, locals::find_locals};
-use tree_sitter_all_languages::parsers::BundledParser;
+use tree_sitter_all_languages::BundledParser;
 use walkdir::WalkDir;
 
 // TODO: Could probably add some filters here for managing/enabling/disabling
