+++ docker-images/syntax-highlighter/crates/syntax-analysis/src/ctags.rs
@@ -9,7 +9,7 @@ use anyhow::{Context, Result};
 use itertools::intersperse;
 use scip::types::{descriptor::Suffix, symbol_information, Descriptor};
 use serde::{Deserialize, Serialize};
-use tree_sitter_all_languages::parsers::BundledParser;
+use tree_sitter_all_languages::BundledParser;
 
 use crate::{get_globals, globals::Scope};
