+++ docker-images/syntax-highlighter/crates/syntax-analysis/src/highlighting/syntect_scip.rs
@@ -534,10 +534,8 @@ mod test {
         io::Read,
     };
 
+    use crate::snapshot::{dump_document_with_config, EmitSymbol, EmitSyntax, SnapshotOptions};
     use pretty_assertions::assert_eq;
-    use crate::snapshot::{
-        dump_document_with_config, EmitSymbol, EmitSyntax, SnapshotOptions,
-    };
 
     fn snapshot_sciptect_documents(doc: &Document, source: &str) -> String {
         dump_document_with_config(
