+++ docker-images/syntax-highlighter/crates/syntax-analysis/src/highlighting/tree_sitter.rs
@@ -1,19 +1,198 @@
 use anyhow::Result;
+use paste::paste;
 use protobuf::Message;
 use rocket::serde::json::{serde_json::json, Value as JsonValue};
 use scip::types::{Document, Occurrence, SyntaxKind};
-use crate::range::PackedRange;
-use tree_sitter_all_languages::highlights::{
-    get_highlighting_configuration, get_syntax_kind_for_hl,
-};
+use std::collections::HashMap;
+use tree_sitter_all_languages::BundledParser;
 use tree_sitter_highlight::{
     Error, Highlight, HighlightConfiguration, HighlightEvent, Highlighter as TSHighlighter,
 };
 
 use crate::highlighting::SourcegraphQuery;
+use crate::range::PackedRange;
+
+macro_rules! include_scip_query {
+    ($lang: expr, $query: literal) => {
+        include_str!(concat!(
+            env!("CARGO_MANIFEST_DIR"),
+            "/queries/",
+            $lang,
+            "/",
+            $query,
+            ".scm"
+        ))
+    };
+}
+pub(crate) use include_scip_query;
+
+#[rustfmt::skip]
+// Table of (@CaptureGroup, SyntaxKind) mapping.
+//
+// Any capture defined in a query will be mapped to the following SyntaxKind via the highlighter.
+//
+// To extend what types of captures are included, simply add a line below that takes a particular
+// match group that you're interested in and map it to a new SyntaxKind.
+//
+// We can also define our own new capture types that we want to use and add to queries to provide
+// particular highlights if necessary.
+//
+// (I can also add per-language mappings for these if we want, but you could also just do that with
+//  unique match groups. For example `@rust-bracket`, or similar. That doesn't need any
+//  particularly new rust code to be written. You can just modify queries for that)
+const MATCHES_TO_SYNTAX_KINDS: &[(&str, SyntaxKind)] = &[
+    ("boolean",                 SyntaxKind::BooleanLiteral),
+    ("character",               SyntaxKind::CharacterLiteral),
+    ("comment",                 SyntaxKind::Comment),
+    ("conditional",             SyntaxKind::IdentifierKeyword),
+    ("constant",                SyntaxKind::IdentifierConstant),
+    ("identifier.constant",     SyntaxKind::IdentifierConstant),
+    ("constant.builtin",        SyntaxKind::IdentifierBuiltin),
+    ("constant.null",           SyntaxKind::IdentifierNull),
+    ("float",                   SyntaxKind::NumericLiteral),
+    ("function",                SyntaxKind::IdentifierFunction),
+    ("method",                  SyntaxKind::IdentifierFunction),
+    ("identifier.function",     SyntaxKind::IdentifierFunction),
+    ("function.builtin",        SyntaxKind::IdentifierBuiltin),
+    ("identifier.builtin",      SyntaxKind::IdentifierBuiltin),
+    ("identifier",              SyntaxKind::Identifier),
+    ("identifier.attribute",    SyntaxKind::IdentifierAttribute),
+    ("tag.attribute",           SyntaxKind::TagAttribute),
+    ("include",                 SyntaxKind::IdentifierKeyword),
+    ("keyword",                 SyntaxKind::IdentifierKeyword),
+    ("keyword.function",        SyntaxKind::IdentifierKeyword),
+    ("keyword.return",          SyntaxKind::IdentifierKeyword),
+    ("number",                  SyntaxKind::NumericLiteral),
+    ("operator",                SyntaxKind::IdentifierOperator),
+    ("identifier.operator",     SyntaxKind::IdentifierOperator),
+    ("property",                SyntaxKind::Identifier),
+    ("punctuation",             SyntaxKind::UnspecifiedSyntaxKind),
+    ("punctuation.bracket",     SyntaxKind::UnspecifiedSyntaxKind),
+    ("punctuation.delimiter",   SyntaxKind::PunctuationDelimiter),
+    ("string",                  SyntaxKind::StringLiteral),
+    ("string.special",          SyntaxKind::StringLiteral),
+    ("string.escape",           SyntaxKind::StringLiteralEscape),
+    ("tag",                     SyntaxKind::UnspecifiedSyntaxKind),
+    ("type",                    SyntaxKind::IdentifierType),
+    ("identifier.type",         SyntaxKind::IdentifierType),
+    ("type.builtin",            SyntaxKind::IdentifierBuiltinType),
+    ("regex.delimiter",         SyntaxKind::RegexDelimiter),
+    ("regex.join",              SyntaxKind::RegexJoin),
+    ("regex.escape",            SyntaxKind::RegexEscape),
+    ("regex.repeated",          SyntaxKind::RegexRepeated),
+    ("regex.wildcard",          SyntaxKind::RegexWildcard),
+    ("identifier",              SyntaxKind::Identifier),
+    ("variable",                SyntaxKind::Identifier),
+    ("identifier.builtin",      SyntaxKind::IdentifierBuiltin),
+    ("variable.builtin",        SyntaxKind::IdentifierBuiltin),
+    ("identifier.parameter",    SyntaxKind::IdentifierParameter),
+    ("variable.parameter",      SyntaxKind::IdentifierParameter),
+    ("identifier.module",       SyntaxKind::IdentifierModule),
+    ("variable.module",         SyntaxKind::IdentifierModule),
+];
+
+/// Add a language highlight configuration to the CONFIGURATIONS global.
+///
+/// This makes it so you don't have to understand how configurations are added,
+/// just add the name of filetype that you want.
+macro_rules! create_configurations {
+    ( $(($name:tt,$dirname:literal)),* ) => {{
+        use tree_sitter_all_languages::BundledParser;
+
+        let mut m = HashMap::new();
+        let highlight_names = MATCHES_TO_SYNTAX_KINDS.iter().map(|hl| hl.0).collect::<Vec<&str>>();
+
+        $(
+            {
+                // Create HighlightConfiguration language
+                let mut lang = HighlightConfiguration::new(
+                    paste! { BundledParser::$name.get_language() },
+                    include_scip_query!($dirname, "highlights"),
+                    include_scip_query!($dirname, "injections"),
+                    include_scip_query!($dirname, "locals"),
+                ).expect(stringify!("parser for '{}' must be compiled", $name));
+
+                // Associate highlights with configuration
+                lang.configure(&highlight_names);
+
+                // Insert into configurations, so we only create once at startup.
+                m.insert(BundledParser::$name, lang);
+            }
+        )*
+
+        {
+            let highlights = vec![
+                include_scip_query!("typescript", "highlights"),
+                include_scip_query!("javascript", "highlights"),
+            ];
+            let mut lang = HighlightConfiguration::new(
+                BundledParser::Typescript.get_language(),
+                &highlights.join("\n"),
+                include_scip_query!("typescript", "injections"),
+                include_scip_query!("typescript", "locals"),
+            ).expect("parser for 'typescript' must be compiled");
+            lang.configure(&highlight_names);
+            m.insert(BundledParser::Typescript, lang);
+        }
+        {
+            let highlights = vec![
+                include_scip_query!("tsx", "highlights"),
+                include_scip_query!("typescript", "highlights"),
+                include_scip_query!("javascript", "highlights"),
+            ];
+            let mut lang = HighlightConfiguration::new(
+                BundledParser::Tsx.get_language(),
+                &highlights.join("\n"),
+                include_scip_query!("tsx", "injections"),
+                include_scip_query!("tsx", "locals"),
+            ).expect("parser for 'tsx' must be compiled");
+            lang.configure(&highlight_names);
+            m.insert(BundledParser::Tsx, lang);
+        }
+
+        m
+    }}
+}
+
+lazy_static::lazy_static! {
+    pub static ref CONFIGURATIONS: HashMap<BundledParser, HighlightConfiguration> = {
+        // NOTE: typescript/tsx crates are included, even though not listed below.
+
+        // You can add any new crate::parsers::Parser variants here.
+        create_configurations!(
+            (C, "c"),
+            (Cpp, "cpp"),
+            (C_Sharp, "c_sharp"),
+            (Go, "go"),
+            (Java, "java"),
+            (Javascript, "javascript"),
+            (Jsonnet, "jsonnet"),
+            (Kotlin, "kotlin"),
+            (Matlab, "matlab"),
+            (Nickel, "nickel"),
+            (Perl, "perl"),
+            (Pod, "pod"),
+            (Python, "python"),
+            (Ruby, "ruby"),
+            (Rust, "rust"),
+            (Scala, "scala"),
+            (Sql, "sql"),
+            (Xlsg, "xlsg"),
+            (Zig, "zig")
+        )
+    };
+}
+
+fn get_highlighting_configuration(filetype: &str) -> Option<&'static HighlightConfiguration> {
+    BundledParser::get_parser(filetype).and_then(|parser| CONFIGURATIONS.get(&parser))
+}
+
+fn get_syntax_kind_for_hl(hl: Highlight) -> SyntaxKind {
+    MATCHES_TO_SYNTAX_KINDS[hl.0].1
+}
 
 // Handle special cases where syntect language names don't match treesitter names.
-pub fn treesitter_language(syntect_language: &str) -> &str {
+pub(crate) fn treesitter_language(syntect_language: &str) -> &str {
     match syntect_language {
         "c++" => "cpp",
         _ => syntect_language,
@@ -80,11 +259,10 @@ pub fn index_language_with_config(
     doc.occurrences.sort_by_key(|a| (a.range[0], a.range[1]));
 
     if include_locals {
-        let parser = tree_sitter_all_languages::parsers::BundledParser::get_parser(filetype);
+        let parser = tree_sitter_all_languages::BundledParser::get_parser(filetype);
         if let Some(parser) = parser {
             // TODO: Could probably write this in a much better way.
-            let mut local_occs =
-                crate::get_locals(parser, code.as_bytes()).unwrap_or_default();
+            let mut local_occs = crate::get_locals(parser, code.as_bytes()).unwrap_or_default();
 
             // Get ranges in reverse order, because we're going to pop off the back of the list.
             //  (that's why we're sorting the opposite way of the document occurrences above).
@@ -200,7 +378,7 @@ impl OffsetManager {
 }
 
 /// Converts a general-purpose syntax highlighting iterator into a sequence of lines of HTML.
-pub struct ScipEmitter {}
+pub(crate) struct ScipEmitter {}
 
 /// Our version of `tree_sitter_highlight::HtmlRenderer`, which emits stuff as a table.
 ///
@@ -260,9 +438,9 @@ mod test {
         io::Read,
     };
 
-    use crate::snapshot::{self, dump_document_with_config};
     use super::*;
     use crate::highlighting::determine_filetype;
+    use crate::snapshot::{self, dump_document_with_config};
 
     fn snapshot_treesitter_syntax_kinds(doc: &Document, source: &str) -> String {
         dump_document_with_config(
@@ -326,7 +504,11 @@ SELECT * FROM my_table
     #[test]
     fn test_all_files() -> Result<(), std::io::Error> {
         let crate_root: std::path::PathBuf = std::env::var("CARGO_MANIFEST_DIR").unwrap().into();
-        let input_dir = crate_root.join("src").join("highlighting").join("snapshots").join("files");
+        let input_dir = crate_root
+            .join("src")
+            .join("highlighting")
+            .join("snapshots")
+            .join("files");
         let dir = read_dir(&input_dir).unwrap();
 
         let mut failed_tests = vec![];
